Move profiles to xtraplatform-features
Move the profiles and their associated property transformations to xtraplatform-features. All aspects related to API building block stays in ldproxy.

A new transformer was introduced to change the type of a codelist property (which may be an integer) to string to correctly represent the value type. Otherwise the value is not encoded in the vector tile.

The profiles to be applied are declared in the tileset of a tile provider as an array with the key `profiles`.

Related to https://github.com/ldproxy/ldproxy/issues/1328.