add DefaultRolesResolver
If no primary geometry is explicitly set, the first spatial property is assigned this role. If no primary instant or interval is explicitly set, the first temporal property is assigned this role.

These defaults are explicitly set in the hydration step. This has to be done after the schema is complete. The new DefaultRolesResolver is executed as the final resolver.

Note that other roles including the primary interval start or end are only applied, if explicitly set in the provider schema.